+++ programs/solana_fpl/src/state/escrow_account.rs
@@ -7,6 +7,6 @@ pub struct EscrowAccount {
     pub payout_first: u64,
     pub payout_second: u64,
     pub payout_third: u64,
-    pub bet_amount: u128,
-    pub usdc_balance: u64,
+    pub bet_amount: u64,
+    pub usdc_balance: u128,
 }
